requirements.txt.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.7
pytz==2020.1
sqlparse==0.3.1
wincertstore==0.2
Werkzeug==1.0.0
Pillow==7.0.0
django-ckeditor==5.9.0
django-js-asset==1.2.2